Dodano strone z edycja pytan, usunieto security
Wykorzystujace db, poniewaz nie dzialalo

+++ src/main/java/sample/data/jpa/service/QuestionService.java
@@ -9,6 +9,6 @@ public interface QuestionService {
     
     void addQuestion(Question quest);
     Question getOneQuestion(Long id);
-    void deleteQuestion(Long id);
+    void deleteQuestion(int id);
     List<Question> getAllQuestions();
 }
